+++ apps/ai-blog-writer/apps/backend/app/features/url2blog/__init__.py
@@ -1,4 +1,5 @@
-# URL2Blog Feature
-from .routes import router
+"""URL2Blog Feature."""
+
+from .api.router import router
 
 __all__ = ["router"]
