+++ apps/ai-blog-writer/apps/backend/app/features/url2blog/api/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI adapters for URL2Blog."""
