+++ apps/ai-blog-writer/apps/backend/app/features/url2blog/pipeline_v2/finalize.py
@@ -2,14 +2,22 @@
 
 from typing import Any
 from fastapi.responses import JSONResponse
+from ..dependencies import PipelineDependencies
 from .finalize_payload import _FinalizePayload
 from .finalize_setup import _FinalizeSetup
 
 
 class _FinalizeResponsePhase(_FinalizeSetup, _FinalizePayload):
 
-    def __init__(self, context: dict[str, Any]) -> None:
+    def __init__(
+        self,
+        context: dict[str, Any],
+        dependencies: PipelineDependencies,
+    ) -> None:
         self.context = context
+        self.dependencies = dependencies
+        self.llm = dependencies.llm
+        self.recorder = dependencies.recorder
 
     def run(self) -> JSONResponse:
         self._initialize()
@@ -21,5 +29,8 @@ class _FinalizeResponsePhase(_FinalizeSetup, _FinalizePayload):
         return JSONResponse(self.response_payload)
 
 
-def _pipeline_v2_finalize_response(context: dict[str, Any]) -> JSONResponse:
-    return _FinalizeResponsePhase(context).run()
+def _pipeline_v2_finalize_response(
+    context: dict[str, Any],
+    dependencies: PipelineDependencies,
+) -> JSONResponse:
+    return _FinalizeResponsePhase(context, dependencies).run()
